requirements.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.4
django-cache-url==3.4.5
django-environ==0.11.2
django-environs==0.1.1
djangorestframework==3.15.1
drf-yasg==1.21.7
environs==11.0.0
gunicorn==22.0.0
inflection==0.5.1
marshmallow==3.21.2
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redbaron==0.9.2
rply==0.7.8
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
whitenoise==6.6.0
